requirements.txt
beautifulsoup4==4.12.3
botcity-framework-base==0.4.3
botcity-framework-web==0.8.4
botcity-maestro-sdk==0.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
et-xmlfile==1.1.0
idna==3.7
importlib_metadata==7.1.0
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==1.26.18
webdriver-manager==4.0.1
zipp==3.18.1
